pyproject.toml
[tool.poetry]
name = "cationbot"
version = "1.1.0"
description = "A Bot to manage our server on Discord."
authors = ["Renan Gomes <renan.gomes@live.com>"]
license = "MIT"
repository = "https://github.com/CationHQ/cationbot"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[tool.pytest.ini_options]
minversion = "6.2.3"
addopts = "-s -rf"
testpaths = ["tests"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.1"
pydantic = {extras = ["dotenv"], version = "^1.8.1"}

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
black = "^20.8b1"
ipdb = "^0.13.7"
pytest = "^6.2.3"
pytest-mock = "^3.5.1"
pytest-cov = "^2.11.1"
Faker = "^6.6.3"
coverage = "^5.5"
autoflake = "^1.4"
pylama = "^7.7.1"
pytest-asyncio = "0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
